backend/requirements.txt
asgiref==3.2.5
cachetools==4.1.1
certifi==2020.4.5.1
cffi==1.14.1
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.4
django-allauth==0.41.0
django-cors-headers==3.2.1
django-filter==2.3.0
django-rest-auth==0.9.5
django-storages==1.9.1
djangorestframework==3.11.0
drf-nested-routers==0.91
google-api-core==1.22.0
google-auth==1.20.1
google-cloud-core==1.4.1
google-cloud-storage==1.30.0
google-crc32c==0.1.0
google-resumable-media==0.7.1
googleapis-common-protos==1.52.0
idna==2.9
mysqlclient==1.4.6
oauthlib==3.1.0
Pillow==7.1.2
protobuf==3.12.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
